refactor: extract recurrence and service-year modules from serviceReport
