+++ src/features/home/screens/HomeScreen.tsx
@@ -17,7 +17,8 @@ import XView from '@/components/ui/layout/XView'
 import YearMilestoneCard from '@/components/YearMilestoneCard'
 import moment from 'moment'
 import useDevice from '@/hooks/useDevice'
-import { getMonthsReports, getServiceYearFromDate } from '@/lib/serviceReport'
+import { getMonthsReports } from '@/lib/serviceReport'
+import { getServiceYearFromDate } from '@/lib/serviceYear'
 import WeekStripTeaser from '@/features/service-reports/components/WeekStripTeaser'
 import i18n from '@/lib/locales'
 import Text from '@/components/ui/MyText'
