+++ src/lib/serviceReport.ts
@@ -1,11 +1,6 @@
 import _ from 'lodash'
 import { Publisher } from '@/types/publisher'
 import {
-  DayPlan,
-  MonthlyByWeekdayConfig,
-  RecurringPlan,
-  RecurringPlanFrequencies,
-  RecurringPlanOverride,
   TimeEntry,
   TimeEntriesByYear,
   TimeEntriesByMonth,
@@ -13,17 +8,19 @@ import {
 import { hasCategory, isLdcEntry } from '@/lib/serviceReportCategory'
 import moment from 'moment'
 import { monthCreditMaxMinutes } from '@/constants/serviceReports'
-import { creditCapMinutesFor, getEntryMode } from '@/lib/publisherCapabilities'
-import { logger } from '@/lib/logger'
-import {
-  DEFAULT_START_TIME_IN_MINUTES,
-  momentStoredDate,
-  normalizeDateForStorage,
-} from '@/lib/normalizeDate'
+import { creditCapMinutesFor } from '@/lib/publisherCapabilities'
+import { momentStoredDate } from '@/lib/normalizeDate'
+import { getServiceYearFromDate } from '@/lib/serviceYear'
 
 // Re-exported for backwards compatibility — canonical home is `types/timeEntry`.
-export { RecurringPlanFrequencies }
-export type { MonthlyByWeekdayConfig, RecurringPlan, RecurringPlanOverride }
+// Recurrence helpers that used to live here now live in `@/lib/recurrence`;
+// Service-Year helpers in `@/lib/serviceYear`.
+export { RecurringPlanFrequencies } from '@/types/timeEntry'
+export type {
+  MonthlyByWeekdayConfig,
+  RecurringPlan,
+  RecurringPlanOverride,
+} from '@/types/timeEntry'
 
 export const calculateProgress = ({
   minutes,
@@ -183,27 +180,6 @@ export const adjustedMinutesForSpecificMonth = (
   }
 }
 
-export const totalMinutesForSpecificMonthUpToDayOfMonth = (
-  serviceReports: TimeEntry[],
-  targetDay: number,
-  targetMonth: number,
-  targetYear: number
-): number => {
-  const totalMinutesForMonth = serviceReports
-    .filter((report) => {
-      const m = momentStoredDate(report.date)
-      return (
-        m.month() === targetMonth &&
-        m.year() === targetYear &&
-        m.date() <= targetDay
-      )
-    })
-    .reduce((accumulator, report) => {
-      return accumulator + report.hours * 60 + report.minutes
-    }, 0)
-
-  return totalMinutesForMonth
-}
 export const ldcMinutesForSpecificMonth = (
   monthsReports: TimeEntry[],
   targetMonth: number,
@@ -321,76 +297,6 @@ export const getDaysLeftInCurrentMonth = () => {
   return daysLeftInMonth
 }
 
-export const getTimeAsMinutesForHourglass = (
-  publisher: Publisher,
-  wentOutForMonth: boolean | null,
-  minutes: number | null
-) => {
-  if (getEntryMode(publisher) === 'checkbox') {
-    if (wentOutForMonth) {
-      return 1
-    }
-    return 0
-  }
-  return minutes
-}
-
-export const serviceReportHoursPerMonthToGoal = ({
-  currentDate,
-  goalHours,
-  serviceReports,
-  serviceYear,
-}: {
-  serviceReports: TimeEntriesByYear
-  currentDate: {
-    month: number
-
-    year: number
-  }
-  goalHours: number
-  serviceYear: number
-}) => {
-  const { maxDate } = serviceYearsDateRange(serviceYear)
-  const annualGoalHours = goalHours * 12
-
-  const now = moment().month(currentDate.month).year(currentDate.year)
-
-  const monthReports = getMonthsReports(
-    serviceReports,
-    currentDate.month,
-    currentDate.year
-  )
-
-  const monthsRemainingOffset = !monthReports.length ? 1 : 0
-
-  const actualMonthsRemaining =
-    moment(maxDate).diff(now, 'months') + monthsRemainingOffset
-
-  const monthsRemaining =
-    actualMonthsRemaining === 0 ? 1 : actualMonthsRemaining
-
-  const serviceYearReports = getServiceYearReports(serviceReports, serviceYear)
-  const totalMinutesForServiceYear = getTotalMinutesForServiceYear(
-    serviceYearReports,
-    serviceYear
-  )
-
-  return _.round(
-    (annualGoalHours * 60 - totalMinutesForServiceYear) / 60 / monthsRemaining,
-    1
-  )
-}
-
-export const serviceYearsDateRange = (serviceYear: number) => {
-  const minDate = moment().month(8).year(serviceYear).startOf('month')
-  const maxDate = moment()
-    .month(7)
-    .year(serviceYear + 1)
-    .endOf('month')
-
-  return { minDate, maxDate }
-}
-
 /**
  * Raw standard/credit logged minutes for one calendar month, before any credit
  * cap is applied. Produced by `getServiceYearMonthlyBreakdowns` and consumed by
@@ -486,17 +392,6 @@ export const getTotalMinutesForServiceYear = (
       applyMonthCreditCap(m.standard, m.credit, effectiveCreditLimitMinutes),
     0
   )
-}
-
-export const getServiceYearFromDate = (moment: moment.Moment) => {
-  const month = moment.month()
-  const year = moment.year()
-
-  if (month < 8) {
-    return year - 1
-  }
-
-  return year
 }
 
 // ---------------------------------------------------------------------------
@@ -645,490 +540,6 @@ export const getAvailableEarlierEndYears = (
   return out
 }
 
-// Helper function to check if a date matches a monthly by weekday pattern.
-// Caller is responsible for passing a calendar-day Date that's already been
-// normalized via `normalizeDateForStorage`; this fn reads it via UTC mode.
-const doesDayMatchMonthlyByWeekday = (
-  day: Date,
-  config: MonthlyByWeekdayConfig
-): boolean => {
-  const momentDay = momentStoredDate(day)
-  const dayWeekday = momentDay.day() // 0-6 (Sunday-Saturday)
-
-  // Check if the weekday matches
-  if (dayWeekday !== config.weekday) {
-    return false
-  }
-
-  // Get the first day of the month and find which week of the month this day is in
-  const firstDayOfMonth = momentDay.clone().startOf('month')
-
-  if (config.weekOfMonth === -1) {
-    // Handle "last [weekday] of month"
-    const lastDayOfMonth = momentDay.clone().endOf('month')
-    const daysFromEnd = lastDayOfMonth.date() - momentDay.date()
-    return daysFromEnd < 7 && dayWeekday === config.weekday
-  } else {
-    // Handle "first", "second", "third", "fourth" [weekday] of month
-    // We need to be more precise about which occurrence this is
-    const firstWeekdayOfMonth = firstDayOfMonth.clone()
-    while (firstWeekdayOfMonth.day() !== config.weekday) {
-      firstWeekdayOfMonth.add(1, 'day')
-    }
-
-    // Calculate which occurrence this is (1st, 2nd, 3rd, 4th)
-    const weeksBetween = momentDay.diff(firstWeekdayOfMonth, 'weeks')
-    const occurrence = weeksBetween + 1
-
-    return occurrence === config.weekOfMonth
-  }
-}
-
-export const getPlansIntersectingDay = (
-  day: Date,
-  plans: RecurringPlan[]
-): RecurringPlan[] => {
-  // `day` is whatever the caller had — often a local-mode Date built from
-  // `selectedMonth.clone().date(i+1).toDate()`. Normalize it so it sits on the
-  // same noon-UTC anchor as the stored plan dates, then read everything in
-  // UTC mode. This is what makes recurrence math TZ-stable: in NZDT (+13) a
-  // local-mode `diff` against a noon-UTC startDate would silently round to a
-  // negative day count and miss every weekly occurrence.
-  const normalizedDay = normalizeDateForStorage(day)
-  const momentDay = momentStoredDate(normalizedDay)
-
-  return plans.filter((plan) => {
-    const { startDate, recurrence } = plan
-    const { frequency, interval, endDate } = recurrence
-
-    if (
-      plan.deletedDates?.some((deletedDate) =>
-        momentStoredDate(deletedDate).isSame(momentDay, 'day')
-      )
-    ) {
-      return false
-    }
-
-    const momentStartDate = momentStoredDate(startDate)
-    const momentEndDate = endDate ? momentStoredDate(endDate) : null
-
-    // Calculate the difference in days between the start date and the given day
-    const daysDiff = momentDay.diff(momentStartDate, 'days')
-
-    // Check if the given day falls within the recurrence pattern
-    switch (frequency) {
-      case RecurringPlanFrequencies.WEEKLY:
-        return (
-          daysDiff % (interval * 7) === 0 &&
-          momentDay.isSameOrAfter(momentStartDate) &&
-          (!momentEndDate || momentDay.isSameOrBefore(momentEndDate))
-        )
-      case RecurringPlanFrequencies.BI_WEEKLY:
-        return (
-          daysDiff % (interval * 14) === 0 &&
-          momentDay.isSameOrAfter(momentStartDate) &&
-          (!momentEndDate || momentDay.isSameOrBefore(momentEndDate))
-        )
-      case RecurringPlanFrequencies.MONTHLY:
-        return (
-          momentDay.date() === momentStartDate.date() &&
-          momentDay.isSameOrAfter(momentStartDate) &&
-          (!momentEndDate || momentDay.isSameOrBefore(momentEndDate))
-        )
-      case RecurringPlanFrequencies.MONTHLY_BY_WEEKDAY:
-        return (
-          recurrence.monthlyByWeekdayConfig &&
-          doesDayMatchMonthlyByWeekday(
-            normalizedDay,
-            recurrence.monthlyByWeekdayConfig
-          ) &&
-          momentDay.isSameOrAfter(momentStartDate) &&
-          (!momentEndDate || momentDay.isSameOrBefore(momentEndDate))
-        )
-      default:
-        return false
-    }
-  })
-}
-
-/**
- * Gets the effective minutes for a recurring plan on a specific date,
- * accounting for overrides. This is the function to use for all calculations
- * that need to display or calculate with the correct planned minutes.
- */
-export const getEffectiveMinutesForRecurringPlan = (
-  plan: RecurringPlan,
-  date: Date
-): number => {
-  const targetDay = momentStoredDate(normalizeDateForStorage(date))
-  const override = plan.overrides?.find((o) =>
-    momentStoredDate(o.date).isSame(targetDay, 'day')
-  )
-  return override ? override.minutes : plan.minutes
-}
-
-/**
- * Gets the effective note for a recurring plan on a specific date, accounting
- * for overrides.
- */
-export const getEffectiveNoteForRecurringPlan = (
-  plan: RecurringPlan,
-  date: Date
-): string | undefined => {
-  const targetDay = momentStoredDate(normalizeDateForStorage(date))
-  const override = plan.overrides?.find((o) =>
-    momentStoredDate(o.date).isSame(targetDay, 'day')
-  )
-  return override?.note || plan.note
-}
-
-/**
- * Gets the effective start time (minutes since midnight) for a recurring plan
- * on a specific date, accounting for overrides. Falls back to noon (720) when
- * neither the override nor the plan has a stored time.
- */
-export const getEffectiveStartTimeInMinutesForRecurringPlan = (
-  plan: RecurringPlan,
-  date: Date
-): number => {
-  const targetDay = momentStoredDate(normalizeDateForStorage(date))
-  const override = plan.overrides?.find((o) =>
-    momentStoredDate(o.date).isSame(targetDay, 'day')
-  )
-  return (
-    override?.startTimeInMinutes ??
-    plan.startTimeInMinutes ??
-    DEFAULT_START_TIME_IN_MINUTES
-  )
-}
-
-export const plannedMinutesToCurrentDayForMonth = (
-  month: number,
-  year: number,
-  dayPlans: DayPlan[],
-  recurringPlans: RecurringPlan[]
-) => {
-  const selectedMonth = moment().month(month).year(year)
-
-  const dayOfMonth = selectedMonth.isBefore(moment(), 'month')
-    ? selectedMonth.daysInMonth()
-    : moment().date()
-
-  let count = 0
-
-  if (selectedMonth.isAfter(moment(), 'month')) {
-    return 0
-  }
-
-  Array(dayOfMonth)
-    .fill(1)
-    .forEach((_, i) => {
-      const day = selectedMonth.clone().date(i + 1)
-      const dayDate = day.toDate()
-
-      const dayPlan = dayPlans.find((plan) =>
-        momentStoredDate(plan.date).isSame(
-          momentStoredDate(normalizeDateForStorage(dayDate)),
-          'day'
-        )
-      )
-
-      const recurringPlansForDay = getPlansIntersectingDay(
-        dayDate,
-        recurringPlans
-      )
-
-      // Get the highest recurring plan for the day, but use effective minutes (with overrides)
-      const highestRecurringPlanForDay = recurringPlansForDay
-        .map((plan) => ({
-          plan,
-          effectiveMinutes: getEffectiveMinutesForRecurringPlan(plan, dayDate),
-        }))
-        .sort((a, b) => b.effectiveMinutes - a.effectiveMinutes)[0]
-
-      if (dayPlan) {
-        count += dayPlan.minutes
-      } else if (highestRecurringPlanForDay) {
-        count += highestRecurringPlanForDay.effectiveMinutes
-      }
-    })
-
-  return count
-}
-
-/**
- * Optimized calculation for planned minutes up to current day using
- * pre-computed recurring plans.
- */
-export const calculatePlannedMinutesToCurrentDayOptimized = (
-  month: number,
-  year: number,
-  dayPlans: DayPlan[],
-  recurringPlans: RecurringPlan[]
-): number => {
-  const perfStart = performance.now()
-  const selectedMonth = moment().month(month).year(year)
-
-  const dayOfMonth = selectedMonth.isBefore(moment(), 'month')
-    ? selectedMonth.daysInMonth()
-    : moment().date()
-
-  if (selectedMonth.isAfter(moment(), 'month')) {
-    return 0
-  }
-
-  logger.log(
-    `[calculateCurrentDayOptimized] Calculating for ${selectedMonth.format('MMMM YYYY')} up to day ${dayOfMonth} with ${dayPlans.length} day plans and ${recurringPlans.length} recurring plans`
-  )
-
-  // Create day plan lookup map for O(1) access
-  const dayPlanMap = new Map(
-    dayPlans.map((p) => [momentStoredDate(p.date).format('YYYY-MM-DD'), p])
-  )
-  logger.log(
-    `[calculateCurrentDayOptimized] Created day plan map with ${dayPlanMap.size} entries`
-  )
-
-  // Pre-compute recurring plans for the range
-  const minDate = selectedMonth.clone().startOf('month')
-  const maxDate = selectedMonth.clone().date(dayOfMonth)
-  const recurringPlanCache = precomputeRecurringPlansForRange(
-    recurringPlans,
-    minDate,
-    maxDate
-  )
-
-  let count = 0
-
-  for (let i = 0; i < dayOfMonth; i++) {
-    const day = selectedMonth.clone().date(i + 1)
-    const dayKey = day.format('YYYY-MM-DD')
-
-    const dayPlan = dayPlanMap.get(dayKey)
-
-    if (dayPlan) {
-      count += dayPlan.minutes
-    } else {
-      // Get pre-computed recurring plans for this day
-      const recurringPlansForDay = recurringPlanCache.get(dayKey) || []
-
-      // Find the highest minutes value
-      const highestMinutes = recurringPlansForDay.reduce(
-        (max, p) => Math.max(max, p.effectiveMinutes),
-        0
-      )
-
-      count += highestMinutes
-    }
-  }
-
-  logger.log(
-    `[calculateCurrentDayOptimized] Completed in ${(performance.now() - perfStart).toFixed(2)}ms - total minutes: ${count}`
-  )
-
-  return count
-}
-
-/**
- * Generates a stable hash of plans to detect if cache needs invalidation. This
- * is a simple hash based on plan count and last modified times.
- */
-export const generatePlanHash = (
-  dayPlans: DayPlan[],
-  recurringPlans: RecurringPlan[]
-): string => {
-  const dayPlanIds = dayPlans
-    .map((p) => p.id)
-    .sort()
-    .join(',')
-  const recurringPlanIds = recurringPlans
-    .map((p) => p.id)
-    .sort()
-    .join(',')
-  return `d:${dayPlans.length}:${dayPlanIds}|r:${recurringPlans.length}:${recurringPlanIds}`
-}
-
-/**
- * Pre-computes all recurring plan occurrences for a given date range. Returns a
- * Map where key is date string (YYYY-MM-DD) and value is array of plans with
- * effective minutes.
- */
-export const precomputeRecurringPlansForRange = (
-  recurringPlans: RecurringPlan[],
-  minDate: moment.Moment,
-  maxDate: moment.Moment
-): Map<string, { plan: RecurringPlan; effectiveMinutes: number }[]> => {
-  const perfStart = performance.now()
-  const cache = new Map<
-    string,
-    { plan: RecurringPlan; effectiveMinutes: number }[]
-  >()
-
-  logger.log(
-    `[precomputeRecurringPlans] Starting pre-computation for ${recurringPlans.length} plans from ${minDate.format('YYYY-MM-DD')} to ${maxDate.format('YYYY-MM-DD')}`
-  )
-
-  recurringPlans.forEach((plan) => {
-    const current = minDate.clone()
-
-    while (current.isSameOrBefore(maxDate)) {
-      const currentDate = current.toDate()
-      const plansForDay = getPlansIntersectingDay(currentDate, [plan])
-
-      if (plansForDay.length > 0) {
-        const key = current.format('YYYY-MM-DD')
-        const existing = cache.get(key) || []
-        existing.push({
-          plan,
-          effectiveMinutes: getEffectiveMinutesForRecurringPlan(
-            plan,
-            currentDate
-          ),
-        })
-        cache.set(key, existing)
-      }
-
-      current.add(1, 'd')
-    }
-  })
-
-  logger.log(
-    `[precomputeRecurringPlans] Completed in ${(performance.now() - perfStart).toFixed(2)}ms - cached ${cache.size} unique dates`
-  )
-
-  return cache
-}
-
-/**
- * Optimized calculation for monthly planned minutes using pre-computed
- * recurring plans.
- */
-export const calculateMonthlyPlannedMinutesOptimized = (
-  month: number,
-  year: number,
-  dayPlans: DayPlan[],
-  recurringPlans: RecurringPlan[]
-): number => {
-  const perfStart = performance.now()
-  const selectedMonth = moment().month(month).year(year)
-  const dayOfMonth = selectedMonth.daysInMonth()
-
-  logger.log(
-    `[calculateMonthlyOptimized] Calculating for ${selectedMonth.format('MMMM YYYY')} with ${dayPlans.length} day plans and ${recurringPlans.length} recurring plans`
-  )
-
-  // Create day plan lookup map for O(1) access
-  const dayPlanMap = new Map(
-    dayPlans.map((p) => [momentStoredDate(p.date).format('YYYY-MM-DD'), p])
-  )
-  logger.log(
-    `[calculateMonthlyOptimized] Created day plan map with ${dayPlanMap.size} entries`
-  )
-
-  // Pre-compute recurring plans for the month
-  const minDate = selectedMonth.clone().startOf('month')
-  const maxDate = selectedMonth.clone().endOf('month')
-  const recurringPlanCache = precomputeRecurringPlansForRange(
-    recurringPlans,
-    minDate,
-    maxDate
-  )
-
-  let count = 0
-
-  for (let i = 0; i < dayOfMonth; i++) {
-    const day = selectedMonth.clone().date(i + 1)
-    const dayKey = day.format('YYYY-MM-DD')
-
-    const dayPlan = dayPlanMap.get(dayKey)
-
-    if (dayPlan) {
-      count += dayPlan.minutes
-    } else {
-      // Get pre-computed recurring plans for this day
-      const recurringPlansForDay = recurringPlanCache.get(dayKey) || []
-
-      // Find the highest minutes value
-      const highestMinutes = recurringPlansForDay.reduce(
-        (max, p) => Math.max(max, p.effectiveMinutes),
-        0
-      )
-
-      count += highestMinutes
-    }
-  }
-
-  logger.log(
-    `[calculateMonthlyOptimized] Completed in ${(performance.now() - perfStart).toFixed(2)}ms - total minutes: ${count}`
-  )
-
-  return count
-}
-
-/**
- * Optimized calculation for annual planned minutes using pre-computed recurring
- * plans.
- */
-export const calculateAnnualPlannedMinutesOptimized = (
-  serviceYear: number,
-  dayPlans: DayPlan[],
-  recurringPlans: RecurringPlan[]
-): number => {
-  const perfStart = performance.now()
-  const { minDate, maxDate } = serviceYearsDateRange(serviceYear)
-
-  logger.log(
-    `[calculateAnnualOptimized] Calculating service year ${serviceYear} (${minDate.format('YYYY-MM-DD')} to ${maxDate.format('YYYY-MM-DD')}) with ${dayPlans.length} day plans and ${recurringPlans.length} recurring plans`
-  )
-
-  // Create day plan lookup map for O(1) access
-  const dayPlanMap = new Map(
-    dayPlans.map((p) => [momentStoredDate(p.date).format('YYYY-MM-DD'), p])
-  )
-  logger.log(
-    `[calculateAnnualOptimized] Created day plan map with ${dayPlanMap.size} entries`
-  )
-
-  // Pre-compute recurring plans for the entire year
-  const recurringPlanCache = precomputeRecurringPlansForRange(
-    recurringPlans,
-    minDate,
-    maxDate
-  )
-
-  let minutes = 0
-  const current = minDate.clone()
-
-  while (current.isSameOrBefore(maxDate)) {
-    const dayKey = current.format('YYYY-MM-DD')
-
-    const dayPlan = dayPlanMap.get(dayKey)
-
-    if (dayPlan) {
-      minutes += dayPlan.minutes
-    } else {
-      // Get pre-computed recurring plans for this day
-      const recurringPlansForDay = recurringPlanCache.get(dayKey) || []
-
-      // Find the highest minutes value
-      const highestMinutes = recurringPlansForDay.reduce(
-        (max, p) => Math.max(max, p.effectiveMinutes),
-        0
-      )
-
-      minutes += highestMinutes
-    }
-
-    current.add(1, 'd')
-  }
-
-  logger.log(
-    `[calculateAnnualOptimized] Completed in ${(performance.now() - perfStart).toFixed(2)}ms - total minutes: ${minutes}`
-  )
-
-  return minutes
-}
-
 type ReportQueryResult = {
   month: number
   year: number
@@ -1185,31 +596,4 @@ export const getMonthsReports = (
     return []
   }
   return [...year[month]] // Need to return new array so memoization functions doesn't reference existing array
-}
-
-export const getServiceYearReports = (
-  serviceReports: TimeEntriesByYear,
-  serviceYear: number
-): TimeEntriesByYear => {
-  const result: TimeEntriesByYear = {}
-  const first = getYearsReports(serviceReports, serviceYear)
-  const firstYear: TimeEntriesByMonth = {}
-
-  for (let month = 8; month < 12; month++) {
-    if (first[month]) {
-      firstYear[month] = first[month]
-    }
-  }
-  result[serviceYear] = firstYear
-
-  const second = getYearsReports(serviceReports, serviceYear + 1)
-  const secondYear: TimeEntriesByMonth = {}
-  for (let month = 0; month < 8; month++) {
-    if (second[month]) {
-      secondYear[month] = second[month]
-    }
-  }
-  result[serviceYear + 1] = secondYear
-
-  return result
 }
